WIP: Load materials from toml file
The toml loading is a success, and we're now trying to get it to use
those loaded materials in the actual `lit_color` function.

We ran into some lifetime messiness, which is now confined to
`poly.rs`. We need to redesign how it is structured with fresh
eyes next time.

Maybe elliminating `up_tri` and `down_tri`, or not having them
hold color information? We were probably approaching this code
originally in a too-OOP way because it was ported from our Python
version.